requirements.txt
sentence-transformers
numpy
torch
torchvision
torchaudio
pandas
tez
tensorboard